Moved postInit into TokenApi
Although the AuthorizeApi may call database methods to validate, the stuff related
to refresh tokens and api tokens will live in the new TokenApi class so postInit
should live there.
