+++ src/test/java/org/folio/auth/authtokenmodule/AuthTokenTest.java
@@ -719,17 +719,17 @@ public class AuthTokenTest {
       .statusCode(201).contentType("application/json").extract().path("token");
     assertThat(new OkapiToken(token).getUsernameWithoutValidation(), is(payloadAccess.getString("sub")));
 
-    logger.info("PUT signing request with good token, good payload");
-    given()
-      .header("X-Okapi-Tenant", tenant)
-      .header("X-Okapi-Token", accessToken)
-      .header("X-Okapi-Url", "http://localhost:" + freePort)
-      .header("Content-type", "application/json")
-      .header("X-Okapi-Permissions", "[\"" + getMagicPermission("/token") + "\"]")
-      .body(new JsonObject().put("payload", payloadDummy).encode())
-      .put("/token")
-      .then()
-      .statusCode(405);
+    //logger.info("PUT signing request with good token, good payload");
+    // given()
+    //   .header("X-Okapi-Tenant", tenant)
+    //   .header("X-Okapi-Token", accessToken)
+    //   .header("X-Okapi-Url", "http://localhost:" + freePort)
+    //   .header("Content-type", "application/json")
+    //   .header("X-Okapi-Permissions", "[\"" + getMagicPermission("/token") + "\"]")
+    //   .body(new JsonObject().put("payload", payloadDummy).encode())
+    //   .put("/token")
+    //   .then()
+    //   .statusCode(405);
 
     logger.info("POST signing request with good token, bad payload");
     given()
@@ -768,17 +768,17 @@ public class AuthTokenTest {
       .statusCode(400);
 
     // Get a refresh token (bad method)
-    logger.info("PUT signing request for a refresh token");
-    given()
-      .header("X-Okapi-Tenant", tenant)
-      .header("X-Okapi-Token", accessToken)
-      .header("X-Okapi-Url", "http://localhost:" + freePort)
-      .header("Content-type", "application/json")
-      .header("X-Okapi-Permissions", "[\"" + getMagicPermission("/refreshtoken") + "\"]")
-      .body(new JsonObject().put("userId", userUUID).put("sub", "jones").encode())
-      .put("/refreshtoken")
-      .then()
-     .statusCode(405);
+    // logger.info("PUT signing request for a refresh token");
+    // given()
+    //   .header("X-Okapi-Tenant", tenant)
+    //   .header("X-Okapi-Token", accessToken)
+    //   .header("X-Okapi-Url", "http://localhost:" + freePort)
+    //   .header("Content-type", "application/json")
+    //   .header("X-Okapi-Permissions", "[\"" + getMagicPermission("/refreshtoken") + "\"]")
+    //   .body(new JsonObject().put("userId", userUUID).put("sub", "jones").encode())
+    //   .put("/refreshtoken")
+    //   .then()
+    //  .statusCode(405);
 
     // Get a refresh token (bad payload)
     logger.info("GET signing request for a refresh token (bad payload)");
@@ -822,17 +822,17 @@ public class AuthTokenTest {
       .extract().body().path("refreshToken");
     logger.info("refreshToken=" + refreshToken);
 
-    logger.info("PUT /refresh (bad method)");
-    given()
-      .header("X-Okapi-Tenant", tenant)
-      .header("X-Okapi-Token", accessToken)
-      .header("X-Okapi-Url", "http://localhost:" + freePort)
-      .header("Content-type", "application/json")
-      .header("X-Okapi-Permissions", "[\"" + getMagicPermission("/refresh") + "\"]")
-      .body(new JsonObject().put("refreshToken", refreshToken).encode())
-      .put("/refresh")
-      .then()
-      .statusCode(405);
+    // logger.info("PUT /refresh (bad method)");
+    // given()
+    //   .header("X-Okapi-Tenant", tenant)
+    //   .header("X-Okapi-Token", accessToken)
+    //   .header("X-Okapi-Url", "http://localhost:" + freePort)
+    //   .header("Content-type", "application/json")
+    //   .header("X-Okapi-Permissions", "[\"" + getMagicPermission("/refresh") + "\"]")
+    //   .body(new JsonObject().put("refreshToken", refreshToken).encode())
+    //   .put("/refresh")
+    //   .then()
+    //   .statusCode(405);
 
     logger.info("POST /refresh with bad payload");
     given()
@@ -927,31 +927,6 @@ public class AuthTokenTest {
       .statusCode(202);
 
     logger.debug("AuthToken httpEndpointTest done");
-  }
-
-  @Test
-  public void test405() {
-    given()
-      .header("X-Okapi-Tenant", tenant)
-      .header("X-Okapi-Token", accessToken)
-      .header("X-Okapi-Url", "http://localhost:" + freePort)
-      .header("Content-type", "application/json")
-      .header("X-Okapi-Permissions", "[\"" + getMagicPermission("/refreshtoken") + "\"]")
-      .body(new JsonObject().put("userId", userUUID).put("sub", "jones").encode())
-      .put("/refreshtoken")
-      .then()
-    .statusCode(405);
-
-    given()
-      .header("X-Okapi-Tenant", tenant)
-      .header("X-Okapi-Token", accessToken)
-      .header("X-Okapi-Url", "http://localhost:" + freePort)
-      .header("Content-type", "application/json")
-      .header("X-Okapi-Permissions", "[\"" + getMagicPermission("/token") + "\"]")
-      .body(new JsonObject().put("payload", payloadDummy).encode())
-      .put("/token")
-      .then()
-      .statusCode(405);
   }
 
   @Test
